add league_stats class, calculable module, pry to runner, update stat_tracker with methods called from league_stats class
