+++ lib/calculable.rb
@@ -1,0 +1,9 @@
+module Calculable
+
+  def id_team_hash
+    @id_and_team = {}
+    @teams.each do |team|
+      @id_and_team[team[:team_id]] = team[:teamname]
+    end
+  end
+end
